added some basic junit tests, fixed pom, changed a few things in model, controller and repos

+++ src/main/java/com/circleci/demojavaspring/repository/QuoteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.circleci.demojavaspring.model.Quote;
 
-public interface QuoteRepository extends CrudRepository<Quote, Long> {
+public interface QuoteRepository extends CrudRepository<Quote, Integer> {
 
 }
